Fix some playgrounds in SSR
