+++ package.json
@@ -53,6 +53,7 @@
     "remark-gfm": "^3.0.1",
     "solid-devtools": "^0.26.0",
     "solid-js": "1.7.3",
+    "solid-start": "^0.2.26",
     "tsup": "^6.7.0",
     "tsup-preset-solid": "^0.1.8",
     "turbo": "^1.9.1",
