+++ packages/i18n/dev/index.tsx
@@ -1,7 +1,7 @@
 import { ChainedContextApp } from "./chainedContext";
 import { NoContextI18nApp } from "./createChained";
 
-const Root = () => {
+const App = () => {
   return (
     <div>
       <NoContextI18nApp />
@@ -11,4 +11,4 @@ const Root = () => {
   );
 };
 
-render(() => <Root />, document.getElementById("root")!);
+export default App;
