+++ packages/immutable/dev/index.tsx
@@ -3,7 +3,6 @@ import { Component, For, createSignal } from "solid-js";
 import { TransitionGroup } from "solid-transition-group";
 import { createImmutable } from "../src";
 import { PayloadAction, createSlice, configureStore } from "@reduxjs/toolkit";
-import { createStore, reconcile } from "solid-js/store";
 
 type Todo = {
   id: number;
@@ -176,12 +175,11 @@ const Todos: Component<{ slice: typeof todoArray | typeof todoObject }> = props 
   );
 };
 
-render(
-  () => (
+export default function App() {
+  return (
     <>
       <Todos slice={todoArray} />
       <Todos slice={todoObject} />
     </>
-  ),
-  document.getElementById("root")!,
-);
+  );
+}
