+++ packages/transition-group/dev/index.tsx
@@ -3,8 +3,8 @@ import { A, Route, Router, Routes } from "@solidjs/router";
 import ListPage from "./list-page";
 import SwitchPage from "./switch-page";
 
-render(
-  () => (
+function App() {
+  return (
     <Router>
       <nav class="fixed left-2 top-2 flex space-x-4">
         <A class="text-yellow-400" href="/list">
@@ -21,6 +21,7 @@ render(
         </Routes>
       </div>
     </Router>
-  ),
-  document.getElementById("root")!,
-);
+  );
+}
+
+export default App;
