+++ packages/workers/dev/index.tsx
@@ -1,4 +1,4 @@
-import { Component, createSignal } from "solid-js";
+import { Component, createSignal, onMount } from "solid-js";
 
 import { createWorker, createWorkerPool, createSignaledWorker } from "../src/index";
 
@@ -116,4 +116,8 @@ const App: Component = () => {
   );
 };
 
-export default App;
+export default function () {
+  const [mounted, setMounted] = createSignal(false);
+  onMount(() => setMounted(true));
+  return <>{mounted() && <App />}</>;
+}
