+++ site/package.json
@@ -15,7 +15,6 @@
     "@types/mark.js": "^8.11.8",
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.21",
-    "solid-start": "^0.2.26",
     "solid-start-static": "^0.2.26",
     "tailwindcss": "^3.3.1",
     "tailwindcss-dir": "^4.0.0"
